# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personal Task Manager API",
    "description": "CSE 341 Week 04 — Google OAuth login protects task routes. Log in at /auth/google in this browser before testing protected routes. Users and admins can CRUD their own tasks. Only admins can DELETE users.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://task-manager-6t5y.onrender.com",
      "description": "Render"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/auth/google/callback": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Google login callback",
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Show the logged-in user",
        "responses": {
          "200": { "description": "Logged-in user" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Created automatically after Google OAuth login."
      }
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "User is not logged in with Google",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}